setup for landing page
