+++ src/app/(public)/page.tsx
@@ -1,8 +1,26 @@
-import React from "react";
-import Landing from "./landing/page";
+import { SignedIn, SignedOut, UserButton } from "@clerk/nextjs";
+import Link from "next/link";
 
-const Home = () => {
-	return <Landing />;
+const Landing = () => {
+	return (
+		<>
+			<h1>Landing</h1>
+			<SignedOut>
+				<Link href="/sign-in" className="btn">
+					Sign in
+				</Link>
+				<Link href="/sign-up" className="btn">
+					Sign up
+				</Link>
+			</SignedOut>
+			<SignedIn>
+				<Link href="/dashboard" className="btn">
+					Go To Dashboard
+				</Link>
+				<UserButton afterSignOutUrl="/" />
+			</SignedIn>
+		</>
+	);
 };
 
-export default Home;
+export default Landing;
